+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 File format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## v0.2 - Not Released
+## v0.3 - 2024-09-11
 
-Under development using Godot 4.1.1, ivoyager 0.0.16-dev.
+Developed for Godot 4.3, ivoyager_core 0.0.19.dev
+
+## Changed
+* Rings exports separate LOD files to support Godot 4.3.
+* Asteroids converter reads epoch from source data and tests consistency.
+
+
+## v0.2 - 2023-09-25
+
+Developed for Godot 4.1.1, ivoyager 0.0.16-dev.
 
 ## Changed
 * Asteroid binaries restructured for ivoyager 0.0.16-dev.20230925
@@ -18,4 +27,3 @@ Developed for Godot 3.5.2, ivoyager 0.0.14.
 ### Added
 * Exports binaries for use in ivoyager_assets/asteroid_binaries/.
 * Exports Saturn rings texture for use in ivoyager_assets/rings/.
-